prefilled form logic and fixing some styles
