+++ labefoods/src/components/index.js
@@ -2,3 +2,4 @@ export * from './styled-containers';
 export * from './inputs';
 export * from './footer-menu';
 export * from './header';
+export * from './prefilled-form';
