+++ labefoods/src/global/globalState.js
@@ -1,17 +1,7 @@
 import React, { useRef } from "react";
 import { GlobalStateContext } from "./globalStateContext";
-import { useRequestData } from './../hooks/useRequestData';
-import { BASE_URL } from './../constants/constants';
-import { useState } from "react";
-
 
 export function GlobalState(props) {
-
-    // STATES
-
-    // REQUESTS
-
-
 
     // FUNCTIONS
 
@@ -23,7 +13,7 @@ export function GlobalState(props) {
 
     const validateEmail = email => /[a-z0-9_.-]+@[a-z]{3}[.a-z]?/.test(email);
     const validatePassword = password => /.{6,}/.test(password);
-    const validateWords = word => /[a-zA-ZáàâãéèêíïóôõöúçñÁÀÂÃÉÈÍÏÓÔÕÖÚÇÑ ]{3,}/.test(word);
+    const validateWords = word => /[a-zA-ZáàâãéèêíïóôõöúçñÁÀÂÃÉÈÍÏÓÔÕÖÚÇÑ ]{2,}/.test(word);
     const validateCPF = CPF => /[0-9]{3}\.?[0-9]{3}\.?[0-9]{3}\-?[0-9]{2}/.test(CPF);
     const validateNumber = number => /[0-9]+/.test(number);
 
@@ -39,7 +29,7 @@ export function GlobalState(props) {
                     validateWords,
                     validateCPF,
                     validateNumber,
-                    isValidated                    
+                    isValidated
                 }
             }>
             {props.children}
