+++ labefoods/src/pages/edit-user/index.js
@@ -73,7 +73,7 @@ export function EditUserPage() {
                     label={'Nome'}
                     placeHolder={'Nome e sobrenome'}
                     error={!isNameValid}
-                    helperText={'Digite um nome com pelo menos 3 letras.'}
+                    helperText={'Digite um nome com pelo menos 2 letras.'}
                     required={true}
                 />
                 <GenericInput
